Split networkx-based model from models.py into separate network.py
